Split QueryReadMode out
