+++ kpf/spectrograph/QueryReadMode.py
@@ -5,7 +5,6 @@ import ktl
 from kpf.KPFTranslatorFunction import KPFTranslatorFunction
 from kpf import (log, KPFException, FailedPreCondition, FailedPostCondition,
                  FailedToReachDestination, check_input)
-from kpf.spectrograph.QueryReadMode import QueryReadMode
 
 
 class QueryReadMode(KPFTranslatorFunction):
@@ -46,23 +45,6 @@ class QueryReadMode(KPFTranslatorFunction):
         log.debug(msg)
         print(msg)
         return green_mode, red_mode
-
-    @classmethod
-    def post_condition(cls, args, logger, cfg):
-        pass
-
-
-class QueryFastReadMode(KPFTranslatorFunction):
-    '''Returns True if both ACF files are consistent with fast read mode.
-    '''
-    @classmethod
-    def pre_condition(cls, args, logger, cfg):
-        pass
-
-    @classmethod
-    def perform(cls, args, logger, cfg):
-        green_mode, red_mode = QueryReadMode.execute({})
-        return (green_mode == 'fast') and (red_mode == 'fast')
 
     @classmethod
     def post_condition(cls, args, logger, cfg):
